git-svn-id: https://k9copy.svn.sourceforge.net/svnroot/k9copy/trunk@95 250c61c0-cc0d-0410-af45-d4f8fccfe43b

+++ libk9copy/k9dvd.h
@@ -134,7 +134,7 @@ public:
   virtual const bool& geterror();
   virtual const QString& geterrMsg();
   virtual const bool& getopened();
-  virtual float getsizeSelected(bool _streams);
+  virtual uint64_t getsizeSelected(bool _streams);
   virtual QString &getFormat();
   virtual k9DVDTitle* getstart();
   virtual void setstart(k9DVDTitle* title);
